[01571] Relocate remaining Tests folder apps to Widgets/Inputs
- Move NullableInputsApp.cs to Apps/Widgets/Inputs/ (namespace already correct)
- Move ValidationTestApp.cs to Apps/Widgets/Inputs/ValidationApp.cs
  - Update namespace: Tests -> Widgets.Inputs
  - Update group: ["Tests"] -> ["Widgets", "Inputs"]
  - Rename class: ValidationTestApp -> ValidationApp
- Delete empty Tests/ directory
